# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a music subscription service server.",
        "title": "Music Subscription Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/playlists": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all playlists created by the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlists"
                ],
                "summary": "List playlists",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/music-service_internal_features_playlists_model.Playlist"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new playlist for the current user. FREE users can create up to 3 playlists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlists"
                ],
                "summary": "Create a playlist",
                "parameters": [
                    {
                        "description": "Playlist creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_playlists_handler.CreatePlaylistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_features_playlists_model.Playlist"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or missing title",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Playlist limit exceeded)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playlists/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information of a user's playlist by ID. Users can only get their own playlists.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlists"
                ],
                "summary": "Get a playlist by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_features_playlists_model.Playlist"
                        }
                    },
                    "400": {
                        "description": "Invalid playlist ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not owner of the playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Playlist not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the title and description of a user's playlist. Users can only update their own playlists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlists"
                ],
                "summary": "Update a playlist",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated playlist payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_playlists_handler.CreatePlaylistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_features_playlists_model.Playlist"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not owner of the playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Playlist not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a user's playlist by ID. Users can only delete their own playlists.",
                "tags": [
                    "playlists"
                ],
                "summary": "Delete a playlist",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (Playlist successfully deleted)"
                    },
                    "400": {
                        "description": "Invalid playlist ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not owner of the playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Playlist not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/playlists/{playlist_id}/tracks/{track_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a track by ID to a user's playlist by ID. Users can only modify their own playlists.",
                "tags": [
                    "playlists"
                ],
                "summary": "Add a track to a playlist",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Playlist ID",
                        "name": "playlist_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Track ID",
                        "name": "track_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created (Track added to playlist)"
                    },
                    "400": {
                        "description": "Invalid playlist ID or track ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not owner of the playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Playlist not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict (Track already in playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a track by ID from a user's playlist by ID. Users can only modify their own playlists.",
                "tags": [
                    "playlists"
                ],
                "summary": "Remove a track from a playlist",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Playlist ID",
                        "name": "playlist_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Track ID",
                        "name": "track_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (Track successfully removed from playlist)"
                    },
                    "400": {
                        "description": "Invalid playlist ID or track ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not owner of the playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Playlist not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_features_playlists_handler.CreatePlaylistRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "music-service_internal_core_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "music-service_internal_features_playlists_model.Playlist": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "ВАЖНО: вводите с префиксом \"Bearer \". Пример: Bearer eyJhbGci...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}